requirements.txt
# Note: These requirements are also specified in pyproject.toml
# This file is kept for development environment setup and compatibility
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=1.0.3
playwright>=1.30.0
beautifulsoup4>=4.9.3
APScheduler>=3.9.1
SQLAlchemy>=1.4.23
aiohttp>=3.8.1
python-dotenv>=0.19.0
pydantic>=1.8.2
requests>=2.26.0
lxml>=4.9.0
spacy>=3.1.0
aiosqlite~=0.20
numpy>=1.26.0,<3
pillow~=10.4
tf-playwright-stealth>=1.1.0
xxhash~=3.4
rank-bm25~=0.2
aiofiles>=24.1.0
colorama~=0.4
snowballstemmer~=2.2
pyOpenSSL>=24.3.0
psutil>=6.1.1
nltk>=3.9.1
rich>=13.9.4
cssselect>=1.2.0